moved all CausalInference code into extension
