+++ Project.toml
@@ -5,7 +5,6 @@ authors = ["Patrick Altmeyer <p.altmeyer@tudelft.nl> and contributors"]
 
 [deps]
 CategoricalArrays = "324d7699-5711-5eae-9e2f-1d82baa6b597"
-CausalInference = "8e462317-f959-576b-b3c1-403f26cec956"
 ChainRulesCore = "d360d2e6-b24c-11e9-a2a3-2a2ae2dbcce4"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 DifferentiationInterface = "a0c0ee7d-e4b9-4e03-894e-1c5f64a51d63"
@@ -13,7 +12,6 @@ Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 EnergySamplers = "f446124b-5d5e-4171-a6dd-a1d99768d3ce"
 Flux = "587475ba-b771-5e3f-ad9e-33799f191a9c"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
-Graphs = "86223c79-3864-5bf0-83f7-82e725a168b6"
 KernelFunctions = "ec8451be-7e33-11e9-00cf-bbf324bd1392"
 LazyArtifacts = "4af54fe1-eca0-43a8-85a7-787d91b784e3"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
@@ -47,7 +45,6 @@ LaplaceReduxExt = "LaplaceRedux"
 [compat]
 Aqua = "0.8"
 CategoricalArrays = "0.10, 1"
-CausalInference = "0.19.2"
 ChainRulesCore = "1.15"
 DataFrames = "1"
 DecisionTree = "0.12.3, 0.12.4"
@@ -56,7 +53,6 @@ Distributions = "0.25.97"
 EnergySamplers = "1"
 Flux = "0.12, 0.13, 0.14, 0.15, 0.16"
 ForwardDiff = "1.3"
-Graphs = "1.11.1"
 JointEnergyModels = "0.1, 0.2"
 KernelFunctions = "0.10"
 LaplaceRedux = "1"
