+++ src/data_preprocessing/counterfactual_data.jl
@@ -2,10 +2,6 @@ using MLJBase: MLJBase, Continuous, Finite
 using StatsBase: StatsBase, ZScoreTransform
 using Tables: Tables
 
-using Graphs
-
-using CausalInference: CausalInference
-
 """
     InputTransformer
 
@@ -19,7 +15,6 @@ const InputTransformer = Union{
     StatsBase.AbstractDataTransform,
     MultivariateStats.AbstractDimensionalityReduction,
     GenerativeModels.AbstractGenerativeModel,
-    CausalInference.SCM,
 }
 
 """
@@ -31,7 +26,6 @@ const TypedInputTransformer = Union{
     Type{<:StatsBase.AbstractDataTransform},
     Type{<:MultivariateStats.AbstractDimensionalityReduction},
     Type{<:GenerativeModels.AbstractGenerativeModel},
-    Type{<:CausalInference.SCM},
 }
 
 """
@@ -289,20 +283,4 @@ function transformable_features(
     ]
     # Returns indices of columns that have varying values:
     return counterfactual_data.features_continuous[idx_not_all_equal]
-end
-
-"""
-    transformable_features(
-        counterfactual_data::CounterfactualData, input_encoder::Type{CausalInference.SCM}
-    )
-
-Returns the indices of all features that have causal parents.
-"""
-function transformable_features(
-    counterfactual_data::CounterfactualData, input_encoder::Type{CausalInference.SCM}
-)
-    # Find all nodes that have causal parents
-    g = counterfactual_data.input_encoder.dag
-    child_causal_nodes = [v for v in vertices(g) if indegree(g, v) >= 1]
-    return child_causal_nodes
 end
