+++ src/data_preprocessing/transformer.jl
@@ -1,5 +1,3 @@
-using CausalInference: CausalInference
-
 """
     fit_transformer!(
         data::CounterfactualData,
@@ -95,23 +93,4 @@ function fit_transformer(
     X = data.X
     dt = GenerativeModels._fit(input_encoder, X; kwargs...)
     return dt
-end
-
-"""
-    fit_transformer(
-        data::CounterfactualData,
-        input_encoder::Type{<:CausalInference.SCM};
-        kwargs...,
-    )
-
-Fit a transformer to the data for a `SCM` object.
-"""
-function fit_transformer(
-    data::CounterfactualData, input_encoder::Type{<:CausalInference.SCM}; kwargs...
-)
-    t = Tables.table(transpose(data.X))
-    est_g, score = CausalInference.ges(t; penalty=1.0, parallel=true)
-    est_dag = CausalInference.pdag2dag!(est_g)
-    scm = CausalInference.estimate_equations(t, est_dag)
-    return scm
 end
